# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/period": {
            "get": {
                "description": "Возращает список всех активных периодов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Получение списка периодов",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Название периода",
                        "name": "searchName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    }
                }
            }
        },
        "/period/create": {
            "post": {
                "description": "Создает новый период с предоставленными данными",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Создание нового периода",
                "parameters": [
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Имя периода",
                        "name": "searchName",
                        "in": "query"
                    },
                    {
                        "description": "Пользовательский объект в формате JSON",
                        "name": "period",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PeriodRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    }
                }
            }
        },
        "/period/{id_period}": {
            "get": {
                "description": "Возвращает информацию о периоде по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Получение периода по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о периоде",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    }
                }
            }
        },
        "/period/{id_period}/delete": {
            "delete": {
                "description": "Удаляет период по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Удаление периода",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Имя периода",
                        "name": "searchName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    }
                }
            }
        },
        "/period/{id_period}/fossil": {
            "post": {
                "description": "Добавляет период к доставке по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Добавление периода к доставке",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Имя периода",
                        "name": "searchName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    }
                }
            }
        },
        "/period/{id_period}/fossil/delete": {
            "delete": {
                "description": "Удаляет период из доставки по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Удаление периода из доставки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "email",
                        "description": "Имя периода",
                        "name": "searchName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список периодов",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.PeriodGetResponse"
                        }
                    }
                }
            }
        },
        "/period/{id_period}/image": {
            "post": {
                "description": "Добавляет изображение к периоду по его ID",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Добавление изображения к периоду",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение периода",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    }
                }
            }
        },
        "/period/{id_period}/update": {
            "put": {
                "description": "Обновляет информацию о периоде по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Период"
                ],
                "summary": "Обновление информации о периоде",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID периода",
                        "name": "id_period",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о периоде",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.Period"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Period": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id_period": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "photo": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.PeriodGetResponse": {
            "type": "object",
            "properties": {
                "id_fossil": {
                    "type": "integer",
                    "example": 1
                },
                "periods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Period"
                    }
                }
            }
        },
        "model.PeriodRequest": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}